chg: move test scripts to test folder.
